Move code to domdec module folder
It's not a module yet, but here's where it will be. Updated
suppressions for module cyclic dependency warnings for domdec and
mdlib.
