+++ docs/doxygen/cycle-suppressions.txt
@@ -1,8 +1,16 @@
+# Order (or more generally, edge selection) is significant (see gmxtree.md);
+# "moduleA -> moduleB" means that moduleA should not depend on moduleB, and is
+# a problem to be addressed at some point.
+fileio -> domdec
+domdec -> imd
+domdec -> ewald
+domdec -> mdlib
+domdec -> pulling
+domdec -> swap
 mdlib -> essentialdynamics
 mdlib -> imd
 mdlib -> ewald
 mdlib -> pulling
-mdlib -> swap
 legacyheaders -> swap
 legacyheaders -> fileio
 timing -> legacyheaders
