+++ src/gromacs/CMakeLists.txt
@@ -94,6 +94,7 @@ add_subdirectory(mdlib)
 add_subdirectory(gmxpreprocess)
 add_subdirectory(listed-forces)
 add_subdirectory(commandline)
+add_subdirectory(domdec)
 add_subdirectory(ewald)
 add_subdirectory(fft)
 add_subdirectory(linearalgebra)
