+++ src/gromacs/mdlib/genborn.c
@@ -42,8 +42,8 @@
 #include <math.h>
 #include <string.h>
 
+#include "gromacs/domdec/domdec.h"
 #include "gromacs/fileio/pdbio.h"
-#include "gromacs/legacyheaders/domdec.h"
 #include "gromacs/legacyheaders/names.h"
 #include "gromacs/legacyheaders/network.h"
 #include "gromacs/legacyheaders/nrnb.h"
