+++ src/gromacs/mdlib/md_support.cpp
@@ -41,7 +41,7 @@
 
 #include <algorithm>
 
-#include "gromacs/legacyheaders/domdec.h"
+#include "gromacs/domdec/domdec.h"
 #include "gromacs/legacyheaders/macros.h"
 #include "gromacs/legacyheaders/md_logging.h"
 #include "gromacs/legacyheaders/mdrun.h"
