+++ src/gromacs/mdlib/ns.c
@@ -42,7 +42,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "gromacs/legacyheaders/domdec.h"
+#include "gromacs/domdec/domdec.h"
 #include "gromacs/legacyheaders/force.h"
 #include "gromacs/legacyheaders/macros.h"
 #include "gromacs/legacyheaders/names.h"
