+++ src/gromacs/mdlib/nsgrid.c
@@ -42,8 +42,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+#include "gromacs/domdec/domdec.h"
 #include "gromacs/fileio/pdbio.h"
-#include "gromacs/legacyheaders/domdec.h"
 #include "gromacs/legacyheaders/macros.h"
 #include "gromacs/legacyheaders/network.h"
 #include "gromacs/legacyheaders/typedefs.h"
